requirements.txt
aliyun_python_sdk_kms==2.14.0
attrs==20.3.0
cryptography==3.2.1
asn1crypto==1.4.0
six==1.15.0
aliyun_python_sdk_core==2.13.30
